requirements.txt
mysql-connector-python
cshedule